fix(image-compression): use different encoder based on platform and use a long value rather than bytes

+++ src/Etdb.ServiceBase.Services.Abstractions/IImageCompressionService.cs
@@ -2,6 +2,6 @@ namespace Etdb.ServiceBase.Services.Abstractions
 {
     public interface IImageCompressionService
     {
-        byte[] Compress(byte[] bytes, string mimeType, byte encodeValue = 75);
+        byte[] Compress(byte[] bytes, string mimeType, long encodeValue = 75);
     }
 }
